fix(example): make sure the message looper is joined after each iteration in multiple-doamin example
Before this commit, the message thread may not exit at the end of each runoff-routing simulation
iteration(see Why). Then the looper can still be active in second iteration and still be waiting
for new messages, which is not expected by us.
In this commit, we force message thread to exit at the end of each iteration via pthread_join()
to solve this issue/bug.

Why: When receiving the last message in message thread at the 1st iteration, it will active the main
thread and turn message thread into inactive status(via pthread_cond_signal). After a while of
running of main thread(main thread may be in second iteration and have alreay created a new message
looper/thread), the old message thread would be picked up by OS for running. But the global message
accumulator of message thread is reset due to creation of new message thread, which would cause the
first message looper to never meet the exit condition.